Add support for Kafka record headers
This adds support for specifying and receiving Kafka record headers
(added in Kafka 0.11) to both the HTTP and gRPC interfaces. For gRPC
headers are simply added as an additional field to the ProdRq and ConsRs
structs - easy. For consuming over HTTP, they're added as an additional
field - also easy.

For producing over HTTP, the HTTP server looks for any HTTP header with
the prefix "X-Kafka-". It strips that prefix to get the Kafka record
header key, and treats the HTTP header value as a base 64 encoded value.