Add config option for the producer partitioner
Both the random and roundrobin partitioners can be useful in situations
where consistent partitioning isn't required. For instance, even though
the default hash partitioner randomly distribute messages with an
unspecified key, its `RequiresConsistency` method returns true,
indicating that it's possible for randomly-partitioned messages to get
published to partitions that are known to be down.

This setting is stored on the configuration struct as a string instead
of a PartitionConstructor, because the test suite expects to be able to
compare values of the config struct, and function pointers are not
comparable.